feat(character-studio): clip state machine, locomotion + foot IK, talk driver
Plan 007 steps 2-4, the core (React-free) motion layer:

- clipStateMachine.ts: base-layer crossfades with per-pair durations
  (idle<->walk 0.25s, walk<->run 0.15s), sit entered/exited through the
  sitDown/standUp transition clips (illegal sit->run routes through
  standUp), additive gesture layer via AnimationUtils.makeClipAdditive
  with a weight envelope (full-body fallback kept behind an option).
  Hips translation tracks rescaled by archetype proportion at
  construction - clips stay authored-once, no retargeting.
- locomotion.ts: circle path (1.2 m default) with eased speed, tangent
  heading, speed->gait mapping with hysteresis, and gait-sync timeScale
  calibrated against the measured clip foot speeds (walk 0.890 m/s,
  run 1.766 m/s from measure_gait_speed).
- footIK.ts: closed-form analytic two-bone solver (knee position from
  the law of cosines in the preserved bend plane) + correction-only
  stance pinning: height+velocity stance detection, 80 ms blend,
  <= 6 cm clamp, animated foot roll preserved.
- talkDriver.ts: amplitude->viseme cells with 60-90 ms anti-flicker
  hold, neutral micro-closes at amplitude dips, optional nod hook, and
  a deterministic speech-shaped amplitude stub (makeSpeechSynthAmplitude).
- proceduralIdle.ts: finish the inherited IdleChannels seam - setChannels
  gates headBob/sway/microTurn writes (restoring base once on disable)
  so Play Mode keeps only breath under clips.
- faceRig.ts: setMouthOverride(cell|null) - the talk layer's seam into
  the drawn mouth without fighting expressions.

35 new tests (machine transitions/weights/gestures/frame-phase order,
IK exactness + clamps + stance + pinning, talk hold/silence/determinism,
locomotion path/hysteresis/timescale, idle channel gating). Also fixes
pre-existing tsc errors in clips.test.ts (gltf-transform arrays).

+++ character-studio/test/core/motion/clips.test.ts
@@ -45,8 +45,8 @@ function channelInfo(anim: Animation) {
     return {
       bone: channel.getTargetNode()?.getName() ?? '',
       path: channel.getTargetPath(),
-      times: Array.from(input),
-      values: Array.from(output),
+      times: Array.from(input as Float32Array),
+      values: Array.from(output as Float32Array),
       components: output.length / input.length,
     }
   })
